+++ crates/karva/tests/it/extensions/fixtures/generators.rs
@@ -5,7 +5,6 @@ use rstest::rstest;
 use crate::common::TestContext;
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_fixture_generator() {
     let test_context = TestContext::with_file(
         "test.py",
@@ -34,7 +33,6 @@ def test_fixture_generator(fixture_generator):
 }
 
 #[rstest]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_fixture_generator_with_second_fixture(#[values("karva", "pytest")] framework: &str) {
     let test_context = TestContext::with_file(
         "test.py",
