+++ crates/karva/tests/it/extensions/tags/expect_fail.rs
@@ -13,7 +13,6 @@ fn get_expect_fail_decorator(framework: &str) -> &str {
 }
 
 #[rstest]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_that_fails(#[values("pytest", "karva")] framework: &str) {
     let context = TestContext::with_file(
         "test.py",
@@ -44,7 +43,6 @@ def test_1():
 }
 
 #[rstest]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_that_passes_karva() {
     let context = TestContext::with_file(
         "test.py",
@@ -84,7 +82,6 @@ def test_1():
 }
 
 #[rstest]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_that_passes_pytest() {
     let context = TestContext::with_file(
         "test.py",
@@ -124,7 +121,6 @@ def test_1():
 }
 
 #[rstest]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_no_reason(#[values("pytest", "karva")] framework: &str) {
     let context = TestContext::with_file(
         "test.py",
@@ -155,7 +151,6 @@ def test_1():
 }
 
 #[rstest]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_with_call(#[values("pytest", "karva")] framework: &str) {
     let context = TestContext::with_file(
         "test.py",
@@ -186,7 +181,6 @@ def test_1():
 }
 
 #[rstest]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_with_true_condition(#[values("pytest", "karva")] framework: &str) {
     let context = TestContext::with_file(
         "test.py",
@@ -217,7 +211,6 @@ def test_1():
 }
 
 #[rstest]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_with_false_condition(#[values("pytest", "karva")] framework: &str) {
     let context = TestContext::with_file(
         "test.py",
@@ -248,7 +241,6 @@ def test_1():
 }
 
 #[rstest]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_with_expression(#[values("pytest", "karva")] framework: &str) {
     let context = TestContext::with_file(
         "test.py",
@@ -280,7 +272,6 @@ def test_1():
 }
 
 #[rstest]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_with_multiple_conditions(#[values("pytest", "karva")] framework: &str) {
     let context = TestContext::with_file(
         "test.py",
@@ -311,7 +302,6 @@ def test_1():
 }
 
 #[rstest]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_with_all_false_conditions(#[values("pytest", "karva")] framework: &str) {
     let context = TestContext::with_file(
         "test.py",
@@ -342,7 +332,6 @@ def test_1():
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_with_single_string_as_reason_karva() {
     let context = TestContext::with_file(
         "test.py",
@@ -368,7 +357,6 @@ def test_1():
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_with_empty_conditions_karva() {
     let context = TestContext::with_file(
         "test.py",
@@ -394,7 +382,6 @@ def test_1():
 }
 
 #[rstest]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_mixed_tests_karva() {
     let context = TestContext::with_file(
         "test.py",
@@ -442,7 +429,6 @@ def test_expected_fail_passes():
 }
 
 #[rstest]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_mixed_tests_pytest() {
     let context = TestContext::with_file(
         "test.py",
@@ -490,7 +476,6 @@ def test_expected_fail_passes():
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_with_runtime_error() {
     let context = TestContext::with_file(
         "test.py",
@@ -516,7 +501,6 @@ def test_1():
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_with_assertion_error() {
     let context = TestContext::with_file(
         "test.py",
@@ -542,7 +526,6 @@ def test_1():
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_with_skip() {
     let context = TestContext::with_file(
         "test.py",
@@ -570,7 +553,6 @@ def test_1():
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_expect_fail_then_unexpected_pass() {
     let context = TestContext::with_file(
         "test.py",
