+++ crates/karva/tests/it/extensions/tags/use_fixtures.rs
@@ -3,7 +3,6 @@ use insta_cmd::assert_cmd_snapshot;
 use crate::common::TestContext;
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_use_fixtures_single_fixture() {
     let test_context = TestContext::with_file(
         "test.py",
@@ -35,7 +34,6 @@ def test_with_use_fixture():
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_use_fixtures_multiple_fixtures() {
     let test_context = TestContext::with_file(
         "test.py",
@@ -71,7 +69,6 @@ def test_with_multiple_use_fixtures():
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_use_fixtures_combined_with_parameter_fixtures() {
     let test_context = TestContext::with_file(
         "test.py",
@@ -106,7 +103,6 @@ def test_combined_fixtures(param_fixture):
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_use_fixtures_with_parametrize() {
     let test_context = TestContext::with_file(
         "test.py",
@@ -141,7 +137,6 @@ def test_use_fixtures_with_parametrize(value):
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_use_fixtures_multiple_decorators() {
     let test_context = TestContext::with_file(
         "test.py",
@@ -179,7 +174,6 @@ def test_multiple_use_fixtures_decorators():
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_use_fixtures_fixture_not_found_but_not_used() {
     let test_context = TestContext::with_file(
         "test.py",
@@ -205,7 +199,6 @@ def test_missing_fixture():
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_use_fixtures_generator_fixture() {
     let test_context = TestContext::with_file(
         "test.py",
@@ -238,7 +231,6 @@ def test_use_fixtures_with_generator():
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_use_fixtures_session_scope() {
     let test_context = TestContext::with_files([(
         "test.py",
@@ -275,7 +267,6 @@ def test_session_2():
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_use_fixtures_mixed_with_normal_fixtures() {
     let test_context = TestContext::with_files([
         (
@@ -317,7 +308,6 @@ def test_mixed_fixtures(shared_fixture):
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_pytest_mark_usefixtures_single_fixture() {
     let test_context = TestContext::with_file(
         "test.py",
@@ -349,7 +339,6 @@ def test_with_pytest_use_fixture():
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_pytest_mark_usefixtures_multiple_fixtures() {
     let test_context = TestContext::with_file(
         "test.py",
@@ -385,7 +374,6 @@ def test_with_multiple_pytest_use_fixtures():
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_pytest_mark_usefixtures_with_parametrize() {
     let test_context = TestContext::with_file(
         "test.py",
@@ -422,7 +410,6 @@ def test_pytest_use_fixtures_with_parametrize(value):
 }
 
 #[test]
-#[ignore = "Will fail unless `maturin build` is ran"]
 fn test_pytest_mark_usefixtures_session_scope() {
     let test_context = TestContext::with_files([(
         "test.py",
